Using archiveNoBaseDir with tar formats fails
This results in an error such as the following:
Error: Command failed: tar: ../build-1.0.0.tgz: Can't add archive to itself
tar: -x: Cannot stat: No such file or directory
tar: ../*: Cannot stat: No such file or directory

The current selector of ".*" was used because the previous selector,
"*", was not including hidden files. However, the * expands the files
from . and ends up including the parent (..) directory, hence the
attempt to exclude it (which works for zip, but is incorrect for
tar commands). However, simply using "." as the file selector yields
the desired results across compression programs.