Jake exceeds stack when a task has many prereqs.
When a task has many prereqs, Jake exceeds the node process stack,
because it calls async callbacks from within a callback, recursively.

Traceback looks something like this:
RangeError: Maximum call stack size exceeded
    at parsePrereqName (jake/lib/task/task.js:44:34)
    at jake/lib/task/task.js:132:24
    at replenish (jake/node_modules/async/lib/async.js:191:21)
    at jake/node_modules/async/lib/async.js:203:33
    at jake/lib/task/task.js:149:13
    at replenish (jake/node_modules/async/lib/async.js:191:21)
    at jake/node_modules/async/lib/async.js:203:33
    at jake/lib/task/task.js:149:13
    at replenish (jake/node_modules/async/lib/async.js:191:21)
    at jake/node_modules/async/lib/async.js:203:33

This fixes the problem by deferring the callback with
async.setImmediate, as described in the section
"Synchronous iteration functions" on this page:
https://caolan.github.io/async/index.html

This also adds a testcase that fails without this fix.