Fix jake -T for directory  tasks
Not sure if exactly `directory` or `packageTask` related, I have both and it breaks when I `jake -T`, all tasks run well still.  

So looks like it needs not to break in the negative case for the regexp, when nothing was matched. This fixes the problem.

Logged `this.action` and `this.action.name` and `this.action.toString()` values that break it.
```
$ jake -T
[Function: action]
{
  name: 'action',
  string: 'function () {\n          jake.mkdirP(name);\n        }'
}
jake aborted.
TypeError: Cannot read properties of undefined (reading 'replace')
    at DirectoryTask._getParams (/Users/falsefalse/.nvm/versions/node/v16.16.0/lib/node_modules/jake/lib/task/task.js:428:10)
```